pyproject.toml
[project]
name = "projeto-ia-entretenimento"
version = "1.0.0"
description = "Projeto starter para criar aplicações de IA para entretenimento"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "google-generative-ai>=1.0.0",
]

#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"
